Conservatively retry journal retrieval
With reusing the journal data apparently doesn't make it to the journal
right away so we can't assume that we can read it back immediately
after journal_send() returns, probably because send buffers are not
flushed immediately.

Hence we retry for about a second until we actually get the data.