sysusers: skip lines with unrecognized type signatures
This introduces a specific error kind to detect and recover from
sysusers parsing failures due to an invalid type signature.
It allows leaving some room for new entry types which could
possibly be introduced upstream in the future.
Thus, lines with unrecognized signatures are now skipped, and a
warning message is logged.