errors: add internal helpers to simplify error handling
This adds a Context trait, as well as implementations of it for Result
and Optional to help consolidate the map_err(|e| format!(...)) logic
that is used to provide propagate error information.  This is inspired
by anyhow, but returns SdError instead to keep the API surface of the
library to a single struct within the library.