sysusers: internal Context helper to simplify error handling
This builds off of the Context trait introduced in #128, and implements
it for the sysusers module. This changes one value, which pulls in
the original error message instead of ignoring it, and adds additional
context on a second parse.  Let me know if you'd rather the error
message for the parsing still be ignored.