fix build in armv7
In armv7-unknown-linux-gnueabihf (ARM 32-bit), FileStat::st_ino is
an u32 instead of an u64, so we cannot directly assign to
JournalStream::inode, which is an u64.

Instead of using u64 in the fields of JournalStream, we use libc::dev_t
and libc::ino_t. The JournalStream::from_metadata function has been removed
since it assumes that the fields are u64 and is no longer used.